+++ src/game/client/c_hairball.cpp
@@ -197,7 +197,7 @@ void C_Hairball::ClientThink()
 	else
 	{
 		// Flip between stopped and starting.
-		if ( fabs( m_flSpinRateX ) > 0.01f )
+		if ( fabsf( m_flSpinRateX ) > 0.01f )
 		{
 			m_flSpinRateX = m_flSpinRateY = 0;
