+++ src/game/client/cstrike/radio_status.cpp
@@ -328,7 +328,7 @@ void CRadioStatus::DrawHeadLabels()
 		// Align it so it never points up or down.
 		Vector vUp( 0, 0, 1 );
 		Vector vRight = CurrentViewRight();
-		if ( fabs( vRight.z ) > 0.95 )	// don't draw it edge-on
+		if ( fabsf( vRight.z ) > 0.95 )	// don't draw it edge-on
 			continue;
 
 		vRight.z = 0;
