+++ src/game/client/dod/c_dod_player.cpp
@@ -92,7 +92,7 @@ public:
 		pParticle->m_flRollDelta += pParticle->m_flRollDelta * ( timeDelta * -2.0f );
 
 		//Cap the minimum roll
-		if ( fabs( pParticle->m_flRollDelta ) < 0.5f )
+		if ( fabsf( pParticle->m_flRollDelta ) < 0.5f )
 		{
 			pParticle->m_flRollDelta = ( pParticle->m_flRollDelta > 0.0f ) ? 0.5f : -0.5f;
 		}
@@ -2451,19 +2451,19 @@ void C_DODPlayer::AvoidPlayers( CUserCmd *pCmd )
 	//Msg( "PRECLAMP: forwardmove=%f, sidemove=%f\n", pCmd->forwardmove, pCmd->sidemove );
 
 	float flForwardScale = 1.0f;
-	if ( pCmd->forwardmove > fabs( cl_forwardspeed.GetFloat() ) )
+	if ( pCmd->forwardmove > fabsf( cl_forwardspeed.GetFloat() ) )
 	{
-		flForwardScale = fabs( cl_forwardspeed.GetFloat() ) / pCmd->forwardmove;
+		flForwardScale = fabsf( cl_forwardspeed.GetFloat() ) / pCmd->forwardmove;
 	}
-	else if ( pCmd->forwardmove < -fabs( cl_backspeed.GetFloat() ) )
+	else if ( pCmd->forwardmove < -fabsf( cl_backspeed.GetFloat() ) )
 	{
-		flForwardScale = fabs( cl_backspeed.GetFloat() ) / fabs( pCmd->forwardmove );
+		flForwardScale = fabsf( cl_backspeed.GetFloat() ) / fabsf( pCmd->forwardmove );
 	}
 
 	float flSideScale = 1.0f;
-	if ( fabs( pCmd->sidemove ) > fabs( cl_sidespeed.GetFloat() ) )
+	if ( fabsf( pCmd->sidemove ) > fabsf( cl_sidespeed.GetFloat() ) )
 	{
-		flSideScale = fabs( cl_sidespeed.GetFloat() ) / fabs( pCmd->sidemove );
+		flSideScale = fabsf( cl_sidespeed.GetFloat() ) / fabsf( pCmd->sidemove );
 	}
 
 	float flScale = min( flForwardScale, flSideScale );
