+++ src/game/client/fx_blood.h
@@ -36,7 +36,7 @@ public:
 
 		//Cap the minimum roll
 		/*
-		if ( fabs( pParticle->m_flRollDelta ) < 0.5f )
+		if ( fabsf( pParticle->m_flRollDelta ) < 0.5f )
 		{
 			pParticle->m_flRollDelta = ( pParticle->m_flRollDelta > 0.0f ) ? 0.5f : -0.5f;
 		}
