+++ src/game/client/portal/c_weapon_physcannon.cpp
@@ -87,7 +87,7 @@ bool C_WeaponPhysCannon::SetupEmitter( void )
 //-----------------------------------------------------------------------------
 static inline void SortAbsVectorComponents( const Vector& src, int* pVecIdx )
 {
-	Vector absVec( fabs(src[0]), fabs(src[1]), fabs(src[2]) );
+	Vector absVec( fabsf(src[0]), fabsf(src[1]), fabsf(src[2]) );
 
 	int maxIdx = (absVec[0] > absVec[1]) ? 0 : 1;
 	if (absVec[2] > absVec[maxIdx])
@@ -169,12 +169,12 @@ void ComputeRenderInfo( mstudiobbox_t *pHitBox, const matrix3x4_t &hitboxToWorld
 	// We project the two longest sides into the vectors perpendicular
 	// to the projection direction, then add in the projection of the perp direction
 	Vector2D size( boxSize[vecIdx[0]], boxSize[vecIdx[1]] );
-	size.x *= fabs( DotProduct( vec[vecIdx[0]], *pXVec ) );
-	size.y *= fabs( DotProduct( vec[vecIdx[1]], *pYVec ) );
+	size.x *= fabsf( DotProduct( vec[vecIdx[0]], *pXVec ) );
+	size.y *= fabsf( DotProduct( vec[vecIdx[1]], *pYVec ) );
 
 	// Add the third component into x and y
-	size.x += boxSize[vecIdx[2]] * fabs( DotProduct( vec[vecIdx[2]], *pXVec ) );
-	size.y += boxSize[vecIdx[2]] * fabs( DotProduct( vec[vecIdx[2]], *pYVec ) );
+	size.x += boxSize[vecIdx[2]] * fabsf( DotProduct( vec[vecIdx[2]], *pXVec ) );
+	size.y += boxSize[vecIdx[2]] * fabsf( DotProduct( vec[vecIdx[2]], *pYVec ) );
 
 	// Bloat a bit, since the shadow wants to extend outside the model a bit
 	size *= 2.0f;
